Fixing Spawning enemies from enemy spawner (desc)
The fact that enemies couldn't spawn from an enemy was due to the fact that the spawner of the enemy spawner was destroyed after it had spawned, the problem was when doing this, the enemies spawned from the enemy spawner no longer had the reference to the spawner.

I therefore created a condition that destroys the spawner after the enemy spawns for all shots other than SpawningShot, in the case of SpawingShot it waits until the enemy spawner is destroyed before destroying the spawner.